requirements.txt
alpaca-py>=0.30.0
yfinance>=0.2.0
ta>=0.11.0
schedule>=1.2.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
Flask>=3.0.0
